requirements.txt
amqp==5.0.1
asgiref~=3.2.10
attrs==20.2.0
auditlog3==1.0.1
autobahn==20.7.1
Automat==20.2.0
Babel==2.8.0
billiard==3.6.3.0
bunch==1.0.1
celery==5.0.1
certifi==2020.6.20
channels==3.0.3
channels-redis==3.3.0
chardet==3.0.4
CodeConvert==3.0.2
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3
daphne==3.0.2
Django==3.1.2
django-ace==1.0.8
django-autoslug==1.9.8
django-celery-beat==2.1.0
django-celery-results==2.0.1
django-constance==2.7.0
django-cors-headers==3.5.0
django-environ==0.4.5
django-extra-fields==3.0.2
django-extensions==3.1.0
django-filter==2.4.0
django-import-export==2.5.0
django-models-ext==1.1.9
django-multiselectfield==0.1.12
django-picklefield==3.0.1
django-redis==4.12.1
django-restql==0.11.0
django-sequences==2.5
django-six==1.0.4
django-timezone-field==4.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
drf-yasg2==1.19.4
fcm-django==1.0.5
firebase-admin==5.0.2
furl==2.1.0
gunicorn==20.0.4
hyperlink==20.0.1
idna==2.10
importlib-metadata==2.0.0
incremental==17.5.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jsonfield==3.1.0
kombu==5.0.2
MarkupSafe==1.1.1
orderedmultidict==1.0.1
packaging==20.4
Pillow==8.0.1
psycopg2-binary==2.8.6
py-money==0.5.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyPEG2==2.15.2
python-crontab==2.5.1
python-dateutil==2.6.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
service-identity==18.1.0
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
stream-chat==2.0.0
TimeConvert==1.5.1
Twisted==20.3.0
txaio==20.4.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.11
vine==5.0.0
whitenoise==5.2.0
xlwt==1.3.0
zipp==3.4.0
zope.interface==5.1.2
django-sslserver==0.22